Create LICENSE
Add Creative Commons Legal Code  License

This commit introduces the Creative Commons Attribution-NonCommercial 4.0 International License to the repository, which applies to the transcribed text content and associated metadata of podcast episodes. This includes URLs to each episode, the relationship with the podcast website, categorizations, and the transcribed content from each broadcast, all of which have been personally collected and transcribed.

The License permits sharing and adapting of the work non-commercially, as long as attribution is given and the derivatives are licensed under the same terms. Commercial use of the materials and data is strictly prohibited.